refactor(council): extract round.rs and synthesis.rs from orchestrator
Phase 0 infrastructure refactor — pure structural extraction with no
behaviour changes.

- Extract run_sequential_round() + run_agent_turn() into round.rs
- Extract run_synthesis() + bridge_synthesis_events() into synthesis.rs
- Introduce RoundContext struct to bundle shared turn dependencies
- Reduce orchestrator.rs to a slim coordinator (~65 lines)
- Update mod.rs module declarations and layout table
- Update README.md MODULE_TABLE with council sub-module entries

All tests pass, clippy clean with -D warnings.
